updated to back down to uniswap v3

+++ scripts/52-ResetUniswapPoolPrice.s.sol
@@ -1123,8 +1123,8 @@ contract ResetUniswapPoolPrice is DeployOrUpgrade {
         int24 maxTick = (currentTick / tickSpacing) * tickSpacing + tickSpacing;
         
         // Ensure ticks are within valid range
-        minTick = minTick < -887270 ? -887270 : minTick;
-        maxTick = maxTick > 887270 ? 887270 : maxTick;
+        minTick = minTick < int24(-887270) ? int24(-887270) : minTick;
+        maxTick = maxTick > int24(887270) ? int24(887270) : maxTick;
         
         // Use minimal amounts
         uint256 minEthAmount = 0.001 ether;
